fix(ts): keep the new agent options off praisonai/mobile's bundle budget
Wiring Agent's `web` and `outputPydantic` options, and re-basing the handoff
errors, pushed praisonai/mobile from 1459.6kB of lazy chunks to 1618.1kB --
past its 1600kB budget -- and put four provider SDKs that are not dependencies
of this package on the graph as bare specifiers a webview cannot resolve. Both
failed the mobile and desktop build gates. Measured against origin/main rather
than guessed: the growth was 62.3kB of zod-to-json-schema plus 179kB of source.

Four changes, each keeping a module off the graph until something asks for it:

- The web-search providers load on demand through a computed specifier, so
  @tavily/ai-sdk and friends are no longer bundled or reported unresolvable.
  `web` now attaches its tool on the first turn instead of in the constructor,
  which the test asserts, with a control that nothing is registered before it.
- zod-to-json-schema and zod are imported through computed specifiers too.
  Neither is a dependency of this package; the existing try/catch already
  treats them as optional, and bundling one cost 62.3kB for a conversion that
  only runs when a caller passes a zod schema.
- CompactionRoute and CompactionStrategy move to context/compaction-types.ts.
  ContextManager needs the strategy names and nothing else, and importing them
  from policy.ts dragged its model tables and token estimators along (24kB).
- The error base (kinds, context protocol, PraisonAIError) moves to
  errors-base.ts, which agent/handoff.ts imports instead of the barrel.

Result: 1526.9kB of 1600kB, no unresolvable imports, and the webview gate
reports OK for every entry. praisonai-ts 2,007 tests and praisonai-mobile
1,372 tests pass; both parity gates still pass.
